agents/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
